GEODE-3345: Refactor ClusterConfigurationDUnitTest to use test rules
* Refactored to use test rules (vs. CliCommandTestBase)
* moved extensions tests into new DUnitTest in geode-core (not wan related).
